task

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -10,7 +10,7 @@
 
 size_t listint_len(const listint_t *h)
 {
-	unsigned int j = 0;
+	size_t j = 0;
 
 	while (h != NULL)
 	{
